backend/requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
joblib
numpy
pandas
requests
python-dotenv
